[Refactor] move CloudConfigurationConstants to shared package (#45134)
