+++ fe/fe-core/src/main/java/com/starrocks/credential/CredentialUtil.java
@@ -17,6 +17,7 @@ package com.starrocks.credential;
 import com.starrocks.catalog.JDBCResource;
 import com.starrocks.connector.iceberg.rest.IcebergRESTCatalog;
 import com.starrocks.connector.odps.OdpsProperties;
+import com.starrocks.connector.share.credential.CloudConfigurationConstants;
 import com.starrocks.credential.azure.AzureCloudConfigurationProvider;
 import com.starrocks.credential.azure.AzureStoragePath;
 import org.apache.logging.log4j.LogManager;
