+++ fe/fe-core/src/main/java/com/starrocks/credential/aliyun/AliyunCloudConfigurationProvider.java
@@ -20,9 +20,9 @@ import com.starrocks.credential.CloudConfigurationProvider;
 
 import java.util.Map;
 
-import static com.starrocks.credential.CloudConfigurationConstants.ALIYUN_OSS_ACCESS_KEY;
-import static com.starrocks.credential.CloudConfigurationConstants.ALIYUN_OSS_ENDPOINT;
-import static com.starrocks.credential.CloudConfigurationConstants.ALIYUN_OSS_SECRET_KEY;
+import static com.starrocks.connector.share.credential.CloudConfigurationConstants.ALIYUN_OSS_ACCESS_KEY;
+import static com.starrocks.connector.share.credential.CloudConfigurationConstants.ALIYUN_OSS_ENDPOINT;
+import static com.starrocks.connector.share.credential.CloudConfigurationConstants.ALIYUN_OSS_SECRET_KEY;
 
 public class AliyunCloudConfigurationProvider implements CloudConfigurationProvider {
