+++ fe/fe-core/src/main/java/com/starrocks/credential/hdfs/HDFSCloudCredential.java
@@ -18,12 +18,11 @@ import autovalue.shaded.com.google.common.common.base.Preconditions;
 import com.staros.proto.FileStoreInfo;
 import com.staros.proto.FileStoreType;
 import com.staros.proto.HDFSFileStoreInfo;
+import com.starrocks.connector.share.credential.CloudConfigurationConstants;
 import com.starrocks.credential.CloudCredential;
 import org.apache.hadoop.conf.Configuration;
 
 import java.util.Map;
-
-import static com.starrocks.credential.CloudConfigurationConstants.HDFS_AUTHENTICATION;
 
 public class HDFSCloudCredential implements CloudCredential {
     public static final String SIMPLE_AUTH = "simple";
@@ -103,7 +102,7 @@ public class HDFSCloudCredential implements CloudCredential {
         fileStore.setFsType(FileStoreType.HDFS);
         HDFSFileStoreInfo.Builder hdfsFileStoreInfo = HDFSFileStoreInfo.newBuilder();
         if (!authentication.isEmpty()) {
-            hdfsFileStoreInfo.putConfiguration(HDFS_AUTHENTICATION, authentication);
+            hdfsFileStoreInfo.putConfiguration(CloudConfigurationConstants.HDFS_AUTHENTICATION, authentication);
             if (authentication.equals(SIMPLE_AUTH) && !userName.isEmpty()) {
                 hdfsFileStoreInfo.setUsername(userName);
             }
