+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/CreateStorageVolumeStmt.java
@@ -16,7 +16,7 @@ package com.starrocks.sql.ast;
 
 import com.google.common.base.Strings;
 import com.starrocks.common.util.PrintableMap;
-import com.starrocks.credential.CloudConfigurationConstants;
+import com.starrocks.connector.share.credential.CloudConfigurationConstants;
 import com.starrocks.sql.parser.NodePosition;
 
 import java.util.List;
