+++ java-extensions/hadoop-ext/pom.xml
@@ -18,6 +18,8 @@
         <maven.compiler.target>8</maven.compiler.target>
         <java-extensions.home>${basedir}/../</java-extensions.home>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <iceberg.version>1.5.0</iceberg.version>
+        <awsv2.version>2.17.257</awsv2.version>
     </properties>
 
     <dependencies>
@@ -30,7 +32,22 @@
         <dependency>
             <groupId>org.apache.iceberg</groupId>
             <artifactId>iceberg-core</artifactId>
-            <version>1.5.0</version>
+            <version>${iceberg.version}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>bundle</artifactId>
+            <version>${awsv2.version}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.iceberg</groupId>
+            <artifactId>iceberg-aws</artifactId>
+            <version>${iceberg.version}</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
